refactor: separate yugabyte / leveldb tests for easier local testing (#1553)
